requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
pydantic==2.7.1
sqlalchemy==2.0.22
psycopg2-binary==2.9.7
alembic==1.11.1
pytest==7.4.2